// plugin-hrm-form/src/components/resources/mappingComponents/uschMappings/types.ts
/**
 * Copyright (C) 2021-2023 Technology Matters
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */

export type UschUiResource = {
  resourceSid: string;
  name: string;
  attributes: {
    alternateName: string;
    address: string;
    phoneNumbers: {
      name: string;
      number: string;
      description?: string;
    }[];
    phoneFax: string;
    emailAddress: string;
    websiteAddress: string;
    description: string;
    categories: string[];
    coverage: string;
    hoursOfOperation: string;
    hoursFormatted: string;
    feeStructure: string;
    comment: string;
    shortDescription: string;
  };
};

export type Language = 'en' | '';